refactor(phases): extract phase1 to discovery.rs
Move the phase1 (Discovery & Cloning) module from phases/mod.rs
to a dedicated discovery.rs file. This reduces mod.rs size and
improves maintainability while preserving the public API via
`pub use discovery as phase1`.
