+++ context/refactor-phases-plan.json
@@ -1,6 +1,6 @@
 {
   "plan_name": "Refactor phases.rs into Directory Module",
-  "last_updated": "2025-11-30",
+  "last_updated": "2025-11-30T00:00:00",
   "goal": "Break up the 5169-line phases.rs into maintainable modules while preserving the public API",
   "session_startup": [
     "Run git status to check branch",
@@ -53,7 +53,7 @@
     {
       "id": "extract-phase1-discovery",
       "name": "Extract phase1 (Discovery & Cloning) to discovery.rs",
-      "status": "pending",
+      "status": "complete",
       "priority": 2,
       "blocked_by": "create-phases-directory",
       "output_file": "src/phases/discovery.rs",
